services/worker_transcribe/requirements.txt
flask
openai-whisper
pydub
numpy
